docs_social/requirements.txt
youtube-transcript-api>=1.2.0
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0 